refactor(AudioSink): 优化代码

+++ src/audiosink/audioformat.hpp
@@ -11,8 +11,8 @@
 
 
 struct AudioDataInfo {
-    qint64 origLength;
-    qint64 processedLength;
+    qint32 origLength;
+    qint32 processedLength;
     qreal speedUpRate;
 };
